Added interfaces, created new controller and renamed
