+++ server/controllers/wazuh-utils/index.ts
@@ -1,0 +1,2 @@
+export * from './wazuh-utils';
+export * from './ui.logs.controller';
